use 127.0.0.1 with postman
using localhost causes a weird name resolution thing: it is resolved as
two different addrs, one of which may be seen as occupied, and one of
which may not. This can cause an issue where attempting to start a
server on an occupied port fails, but the code believes it to have
succeeded.